+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules/
 .idea/
 jquery.min.js
+host/build/
