+++ client/index.html
@@ -7,7 +7,7 @@
     <script type="text/javascript" src="CSInterface.js"></script>
     <script type="text/javascript" src="index.js"></script>
     <!-- Paradigm fail, but needed for endpoint construction -->
-    <script type="text/javascript" src="../host/index.jsx"></script>
+    <script type="text/javascript" src="../host/build/out/index.jsx"></script>
 
     <!-- Change this port, if there are conflicts with other software -->
     <script type="text/javascript">
