+++ host/index.tsx
@@ -5,451 +5,14 @@
 /// <reference path="../typings/extendscript.d.ts"/>
 /// <reference path="../typings/global.d.ts"/>
 
-declare interface Track {
-  overwriteClip(clipProjectItem: ProjectItem, time: Time): void;
-}
-
-class MarkerUtils {
-
-  static getMarkerItemInMarkerFolder(
-    markerColor: string
-  ): undefined | ProjectItem {
-    for (let i = 0; i < app.project.rootItem.children.numItems; i++) {
-      const child = app.project.rootItem.children[i];
-
-      if (child.name === "marker") {
-        for (let j = 0; j < child.children.numItems; j++) {
-          const markerChild = child.children[j];
-
-          if (markerChild.name === markerColor) {
-            return markerChild;
-          }
-        }
-      }
-    }
-    return undefined;
-  }
-
-  static getLastUnnamedMarkerClip(): TrackItem {
-    const currentSequence = app.project.activeSequence;
-    const markerLayer =
-      currentSequence.videoTracks[currentSequence.videoTracks.numTracks - 1];
-    const markerClips = markerLayer.clips;
-    const markerCount = markerClips.numItems;
-
-    const lastMarker = markerClips[markerCount - 1];
-    debugger;
-
-    if (/[0-9]+/.test(lastMarker.name)) {
-      return lastMarker;
-    }
-    return markerClips[markerCount - 2];
-  }
-
-  static addCustomMarker(color: string): void {
-    const currentSequence = app.project.activeSequence;
-    const markerLayer =
-      currentSequence.videoTracks[currentSequence.videoTracks.numTracks - 1];
-
-    Utils.fixPlayHeadPosition();
-    const currentPlayheadPosition = currentSequence.getPlayerPosition();
-
-    const markerChild = MarkerUtils.getMarkerItemInMarkerFolder(color);
-
-    if (markerChild === undefined) {
-      alert("No 'marker' folder found. Please use a viable preset.");
-    } else {
-      markerLayer.overwriteClip(markerChild, currentPlayheadPosition);
-    }
-  }
-
-  static loadMarkersFromCSVFile(): void {
-    const csvFileDialog = File.openDialog("Target CSV File", "*.csv");
-    const csvFile = csvFileDialog.fsName;
-
-    const currentSequence = app.project.activeSequence;
-    const markerLayer =
-      currentSequence.videoTracks[currentSequence.videoTracks.numTracks - 1];
-    const currentPlayheadPosition = currentSequence.getPlayerPosition();
-
-    if (csvFile) {
-      const file = File(csvFile);
-      file.open("r");
-      const fullText = file.read();
-      file.close();
-
-      const lines = fullText.split("\n");
-
-      for (let i = 0; i < lines.length; i++) {
-        const entry = lines[i].split(",");
-
-        if (entry.length === 7) {
-          // Parse csv
-          const seconds =
-            parseInt(entry[0]) * 3600 +
-            parseInt(entry[1]) * 60 +
-            parseInt(entry[2]) +
-            parseInt(entry[3]) / 1000;
-          const mode = entry[4];
-          const message = entry[5];
-          const color = entry[6];
-
-          // Insert clip
-          const markerChild = MarkerUtils.getMarkerItemInMarkerFolder(color);
-          const targetInSeconds = currentPlayheadPosition.seconds + seconds;
-          markerLayer.overwriteClip(markerChild, targetInSeconds);
-
-          // Retrieve clip
-          const clip = MarkerUtils.getLastUnnamedMarkerClip();
-
-          // Set name
-          clip.name = message;
-
-          // Set length
-          if (mode === "mode") {
-            // If mode == mode, get next item with mode mode and calculate length
-            let nextItemSeconds = -1;
-            for (let j = i + 1; j < lines.length; j++) {
-              const newEntry = lines[j].split(",");
-              const nextSeconds =
-                parseInt(newEntry[0]) * 3600 +
-                parseInt(newEntry[1]) * 60 +
-                parseInt(newEntry[2]) +
-                parseInt(newEntry[3]) / 1000;
-              const nextMode = newEntry[4];
-
-              if (nextMode === "mode") {
-                nextItemSeconds = nextSeconds;
-                break;
-              }
-            }
-            if (nextItemSeconds > 0) {
-              const endTime = new Time();
-              endTime.seconds =
-                nextItemSeconds + currentPlayheadPosition.seconds;
-              clip.end = endTime; // not my fault, types problem
-            }
-          }
-        }
-      }
-    }
-  }
-
-  static saveCustomMarkerToTextFile(): void {
-    const currentSequence = app.project.activeSequence;
-    const markerLayer =
-      currentSequence.videoTracks[currentSequence.videoTracks.numTracks - 1];
-    const markerClips = markerLayer.clips;
-    const markerCount = markerClips.numItems;
-    const projectName = app.project.name;
-
-    let output = "Project: " + projectName + "\n";
-    output += "Sequence: " + currentSequence.name + "\n";
-    output += "Marker count: " + markerCount + "\n\n";
-
-    for (let i = 0; i < markerCount; i++) {
-      const clip = markerClips[i];
-      let trueSeconds = parseInt("" + clip.start.seconds);
-
-      const hours = parseInt("" + trueSeconds / 3600);
-      trueSeconds -= hours * 3600;
-
-      const minutes = parseInt("" + trueSeconds / 60);
-      trueSeconds -= minutes * 60;
-
-      const seconds = trueSeconds;
-
-      if (hours > 0) {
-        output += Utils.pad(hours, 2) + ":";
-      }
-      output += Utils.pad(minutes, 2) + ":";
-      output += Utils.pad(seconds, 2) + " - ";
-
-      output += clip.name + "\n";
-    }
-    try {
-      const file = new File();
-      const fileNew = file.saveDlg("Save new file", "*.txt");
-      fileNew.encoding = "UTF8";
-      fileNew.open("w");
-      fileNew.write(output);
-      fileNew.close();
-    } catch (error) {
-      // User hit cancel
-    }
-  }
-
-  static selectCurrentMarker(): void {
-    const clip = MarkerUtils.getCurrentMarkerClip();
-
-    if (clip !== undefined) {
-      clip.setSelected(true, true);
-    }
-  }
-
-  static getCurrentMarkerClip(): TrackItem {
-    const currentSequence = app.project.activeSequence;
-    const markerLayer =
-      currentSequence.videoTracks[currentSequence.videoTracks.numTracks - 1];
-    const markerClips = markerLayer.clips;
-    const markerCount = markerClips.numItems;
-
-    Utils.fixPlayHeadPosition();
-    const currentPlayheadPosition = currentSequence.getPlayerPosition().ticks;
-
-    for (let i = 0; i < markerCount; i++) {
-      const clip = markerClips[i];
-      const startTicks = clip.start.ticks;
-      const endTicks = clip.end.ticks;
-
-      if (
-        parseInt(startTicks) <= parseInt(currentPlayheadPosition) &&
-        parseInt(currentPlayheadPosition) < parseInt(endTicks)
-      ) {
-        return clip;
-      } 
-    }
-  }
-  static deselectAll(): void {
-    const currentSequence = app.project.activeSequence;
-    for (let i = 0; i < currentSequence.videoTracks.numTracks; i++) {
-      const track = currentSequence.videoTracks[i];
-      for (let j = 0; j < track.clips.numItems; j++) {
-        const clip = track.clips[j];
-        clip.setSelected(false, true);
-      }
-    }
-    for (let i = 0; i < currentSequence.audioTracks.numTracks; i++) {
-      const track = currentSequence.videoTracks[i];
-      for (let j = 0; j < track.clips.numItems; j++) {
-        const clip = track.clips[j];
-        clip.setSelected(false, true);
-      }
-    }
-  }
-}
-
-class Utils {
-  static ticksPerSecond = 254016000000;
-
-  static getFirstSelectedClip(videoClip: Boolean) {
-    const currentSequence = app.project.activeSequence;
-    const tracks = videoClip ? currentSequence.videoTracks : currentSequence.audioTracks;
-    for (let i = 0; i < tracks.numTracks; i++) {
-      for (let j = 0; j < tracks[i].clips.numItems; j++) {
-          const currentClip = tracks[i].clips[j];
-          if(currentClip.isSelected()) {
-            return { clip: currentClip,
-              trackIndex: i,
-              clipIndex: j
-            }
-          }
-      }
-    }
-    return null;
-  }
-
-  static getVideoClip(trackIndex: number, clipIndex: number) {
-    const currentSequence = app.project.activeSequence;
-    return currentSequence.videoTracks[trackIndex].clips[clipIndex];
-  }
-
-  static zoomToFit(videoClip) {
-    if(videoClip != null) {
-      const clipSize = this.getClipSize(videoClip.clip);
-      const frameHeight = app.project.activeSequence.frameSizeVertical;
-      const frameWidth = app.project.activeSequence.frameSizeHorizontal;
-
-      const verticalFactor = frameHeight / clipSize.height;
-      const horizontalFactor = frameWidth / clipSize.width;
-      
-      const zoomLevel = Math.max(verticalFactor, horizontalFactor) * 100;
-
-      EffectUtils.setZoomOfCurrentClip(zoomLevel);
-    }
-  }
-
-  static getClipSize(videoClip: TrackItem) {
-    const projectItem = videoClip.projectItem;
-    const videoInfo = Utils.getProjectMetadata(projectItem, 
-      ["Column.Intrinsic.VideoInfo"])[0][0].toString();
-
-    const width = parseInt(videoInfo.split(' ')[0]);
-    const height = parseInt(videoInfo.split(' ')[2]);
-    return {"height": height, "width": width}
-  }
-
-  static getQEVideoClipByStart(trackIndex : number, startInTicks : string) {
-    const currentSequence = qe.project.getActiveSequence();
-    const videoTrack = currentSequence.getVideoTrackAt(trackIndex);
-
-    for(let i = 0; i < videoTrack.numItems; i++) {
-      const clip = videoTrack.getItemAt(i);
-
-      if(clip.start.ticks === startInTicks) {
-        return clip;
-      }
-
-    }
-  }
-
-  static targetAllTracks(target: boolean) {
-    const currentSequence = app.project.activeSequence;
-    for(let i = 0; i < currentSequence.videoTracks.numTracks; i++) {
-      currentSequence.videoTracks[i].setTargeted(target, true)
-    }
-    for(let i = 0; i < currentSequence.audioTracks.numTracks; i++) {
-      currentSequence.audioTracks[i].setTargeted(target, true)
-    }
-  }
-
-  static targetDefaultTracks() { 
-    const currentSequence = app.project.activeSequence;
-    this.targetAllTracks(false);
-    for (let i = 0; i < Math.min(3, currentSequence.videoTracks.numTracks); i++) {
-      currentSequence.videoTracks[i].setTargeted(true, true);
-    }
-    if(currentSequence.audioTracks.numTracks > 0) {
-      currentSequence.audioTracks[0].setTargeted(true, true);
-    }
-  }
-
-  static targetTracks(videoTrack: number, audioTrack: number) {
-    this.targetAllTracks(false);
-
-    const currentSequence = app.project.activeSequence;
-    
-    if(currentSequence.videoTracks.numTracks > videoTrack) {
-      currentSequence.videoTracks[videoTrack].setTargeted(true, true);
-    }
-    if(currentSequence.audioTracks.numTracks > audioTrack) {
-      currentSequence.audioTracks[audioTrack].setTargeted(true, true);
-    }
-  }
-
-  static fixPlayHeadPosition(): void {
-    const currentSequence = app.project.activeSequence;
-    const currentPlayheadPosition = currentSequence.getPlayerPosition().ticks;
-    const ticksPerFrame = currentSequence.getSettings().videoFrameRate.ticks;
-    const newPos = Math.ceil(parseInt(currentPlayheadPosition) / parseInt(ticksPerFrame));
-
-    currentSequence.setPlayerPosition(String(newPos * parseInt(ticksPerFrame)));
-  }
-
-  static pad(num: number, size: number): string {
-    let s = num.toString();
-    while (s.length < size) s = "0" + s;
-    return s;
-  }
-
-  static getProjectItemInRoot(itemName: string): ProjectItem {
-    let projectItem = undefined;
-
-    for (let i = 0; i < app.project.rootItem.children.numItems; i++) {
-      const child = app.project.rootItem.children[i];
-      if (child.name === itemName) {
-        projectItem = child;
-        break;
-      }
-    }
-
-    return projectItem;
-  }
-
-  static getProjectMetadata(projectItem: ProjectItem, fieldNames) {
-    // Based on: https://community.adobe.com/t5/premiere-pro/get-image-size-in-jsx/td-p/10554914?page=1&profile.language=de
-    const kPProPrivateProjectMetadataURI ="http://ns.adobe.com/premierePrivateProjectMetaData/1.0/";
-    if (app.isDocumentOpen()) {
-        if (projectItem) {
-
-            if (ExternalObject.AdobeXMPScript === undefined)
-                ExternalObject.AdobeXMPScript = new ExternalObject("lib:AdobeXMPScript");
-            if (ExternalObject.AdobeXMPScript !== undefined) {
-                let retArray = [];
-                let retArray2 = [];
-                const projectMetadata = projectItem.getProjectMetadata();
-                let xmp = new XMPMeta(projectMetadata);
-                for (let pc = 0; pc < fieldNames.length; pc++) {
-                    if (xmp.doesPropertyExist(kPProPrivateProjectMetadataURI, fieldNames[pc])) {
-                        retArray.push([fieldNames[pc],xmp.getProperty(kPProPrivateProjectMetadataURI, fieldNames[pc])]);
-                        retArray2.push([xmp.getProperty(kPProPrivateProjectMetadataURI,fieldNames[pc])]);
-                    }
-                }
-                return retArray2;
-            }
-        }
-    }
-    return false;
-  }
-}
-
-class EffectUtils {
-
-  static changeAudioLevel(clip: TrackItem, levelInDb: number) {
-    const levelInfo = clip.components[0].properties[1];
-    const level = 20 * Math.log(parseFloat(levelInfo.getValue())) * Math.LOG10E + 15;
-    const newLevel = level + levelInDb;
-    const encodedLevel = Math.min(Math.pow(10, (newLevel - 15)/20), 1.0);
-    levelInfo.setValue(encodedLevel, true);
-  }
-
-  static changeAllAudioLevels(levelInDb: number) {
-    const currentSequence = app.project.activeSequence;
-    for (let i = 0; i < currentSequence.audioTracks.numTracks; i++) {
-      for (let j = 0; j < currentSequence.audioTracks[i].clips.numItems; j++) {
-          const currentClip = currentSequence.audioTracks[i].clips[j];
-          if(currentClip.isSelected()) {
-           this.changeAudioLevel(currentClip, levelInDb);
-          }
-      }
-    }
-  }
-
-  static setZoomOfCurrentClip(zoomLevel: number) {
-    const clipInfo = Utils.getFirstSelectedClip(true)
-    const scaleInfo = clipInfo.clip.components[1].properties[1];
-    scaleInfo.setValue(zoomLevel, true);
-  }
-
-  static applyEffectOnFirstSelectedVideoClip(effectName: String) {
-    const clipInfo = Utils.getFirstSelectedClip(true)
-    const qeClip = Utils.getQEVideoClipByStart(clipInfo.trackIndex, clipInfo.clip.start.ticks)
-    var effect = qe.project.getVideoEffectByName(effectName);
-    qeClip.addVideoEffect(effect);
-
-    // For better usability, always return the newest effects (this ones) properties! 
-    return clipInfo.clip.components[2].properties;
-  }
-
-  static applyDropShadowPreset() { 
-    const shadowEffectProperties = this.applyEffectOnFirstSelectedVideoClip("Schlagschatten");
-    const opacity = shadowEffectProperties[1];
-    const softness = shadowEffectProperties[4];
-
-    opacity.setValue(255, true);
-    softness.setValue(44, true);
-  }
-
-  static applyBlurPreset() {
-    const blurEffectProperties = this.applyEffectOnFirstSelectedVideoClip("Gaußscher Weichzeichner");
-
-    const blurriness = blurEffectProperties[0];
-    const repeatBorderPixels = blurEffectProperties[2];
-
-    blurriness.setValue(42, true);
-    repeatBorderPixels.setValue(true, true);
-  }
-
-  static applyWarpStabilizer() {
-    this.applyEffectOnFirstSelectedVideoClip("Verkrümmungsstabilisierung");
-  }
-
-}
+import { MarkerUtils } from "./MarkerUtils";
+import { Utils } from "./Utils";
+import { EffectUtils } from "./EffectUtils";
 
 /**
  * ALL functions defined here are visible via the localhost REST-like service.
  */
-const host = {
+export const host = {
   /**
    * @swagger
    *
@@ -599,7 +162,7 @@ const host = {
    *          description: Sets the scale of the current clip to 120 percent.
    */
   zoomInTo120percent: function() {
-    EffectUtils.setZoomOfCurrentClip(120);
+    Utils.setZoomOfCurrentClip(120);
   },
 
   /**
@@ -636,7 +199,7 @@ const host = {
 /**
  * These functions are only used internally.
  */
-const framework = {
+export const framework = {
   enableQualityEngineering: function() {
     app.enableQE();
   }
